Initial framework commit
This is just a new laravel application that uses laravel/breeze,
laravel/sail, and laravel/sanctum.

Almost nothing has been modified here except the docker-compose.yml
`platform` option caused an error for the `meilisearch`. This is a
default full text search service that probably isn't even needed for
this application.  For now I just removed the `platform` config
directive so that it would spin up the containers.