Create api routes + resources
This just adds a collection for notes + a resource for a single note to
be serialized/formatted for the API.

I had previously stubbed out the resource object for `Note`, but it
looks like the correct convention is to use a `Resource` suffix, and a
`Collection` suffix in laravel, so I changed it to match that
convention.

Right now this only provides HTTP GET endpoints to pull a single note
and all notes.

By default this will only pull notes that are associated to the
authenticated user.