Add ability to create a note via API
The idea here is to have a single controller per action so that the
controllers can have single responsibility.

The abstract ApiController exists so that we can have an easy way to
return common types of responses (success/error, etc).

To assist with error responses, there is a new ApiException class, which
implements a `render()` method. Laravel uses this behind the scenes to
render the exception as an HTTP response.

To assist the exception there is a DTO just to encapsulate a message +
errors. This might change, but it gets the job done for now.

I'm not sure if this is idiomatic/laravel convention, and if there is a
better way I'm happy to improve it.