Add missing API endpoints
This adds the update, view, and list/index methods.

I decided to make a separate controller for every method, even the
view/index which were previously handled directly in the routes file.

It seems cleaner for all of the endpoints to be structured the same way
instead of having the magic.

This also sneaks in a fix for the previous view endpoint that was not
scoping the query to the current user (oops).

Finally, there is a change in the AppServiceProvider that removes the
default "wrapping" of the json response with `data`. I'm not sure what
the benefit of that is, so it's been removed for now.