Add validation and clean up controllers
This adds some validation:
    - note titles are required and must not be more than 50 characters
    - note bodies (`note`) are not required but must not be more than
      1000 chararacters

I believe the convention in Laravel is to validate request data up
front, rather than at the model level, but it seems safer to apply the
validation just before saving at the model layer as well. This will
better ensure data integrity if the request data does not come from an
HTTP request (a CLI command for example?).

In order to do this, the following interface was added:
`EntityErrorInterface` (there might be a better name?). The idea here is
that any object that implements this interface has to provide a
`getErrors()` and a `getRules()` method.  This has been implemented in
an abstract class (`AbstractModel`) that the tables can extend.

The `AbstractModel` hooks into the `creating` and `updating` model
lifecycle events and checks the data for errors before saving. If it
detects validation errors, it sets them as an `$errors` property and
returns false to prevent the save. The calling code can then call
`getErrors()` to present the validation errors in any context.

Since the convention in Laravel seems to be to create subclasses of
`FormRequest` for defining these rules, that's what I've done. This will
also allow the above pattern to be ditched if it needs to be in factor
of simply validating request data.

There was also a change to the existing migration. The `note` field is
now nullable as it's not required. Normally, this kind of change would
be a separate migration, but there's no existing data or use of this
besides me so the existing migration was simply modified.

The rest of the changes here are minor bug fixes from manually running
the API calls.