Add integration tests
This adds tests for all endpoints:
    - create note
    - update note
    - delete note
    - view notes
    - view note

The `ApiTestCase` just adds a helper method to inject a user into the
request to bypass authentication in the test environment for cases where
we aren't testing the implementation detail of authentication.

The `NoteFactory` might not be needed, but I created it when I was
looking at how the sanctum library handles tests, I noticed that they
used this pattern. The pattern I ended up using is the following:

- Use the laravel helper trait
`Illuminate\Foundation\Testing\DatabaseTransactions`. This just runs
everything in a transaction that gets rolled back. I don't like this as
a test strategy for fixtures/testing actual database writes, but it
works for now.
- Create the data we need as a dependency for each test, e.g. if we need
the existence of another user's note to test that we can't view it, then
create that note in the test.

I ran into some problems with the tests actually connecting to the real
database connection, which I was surprised was possible, so I eventually
used this transaction strategy.  I'm surprised laravel will allow that
behavior by default as it could delete all of your data when running
tests if you're not careful (!)

The rest of the changes here are related to small bugs found while
creating the tests.